requirements.txt
adium-theme-ubuntu==0.3.4
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.0
bs4==0.0.1
cached-property==1.3.0
colorama==0.3.9
cycler==0.10.0
decorator==4.0.6
Django==1.9.5
docker==2.4.2
docker-compose==1.14.0
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
imutils==0.4.3
ipaddress==1.0.18
ipython==2.4.1
Jinja2==2.9.6
jsonschema==2.6.0
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.7.4
mock==2.0.0
networkx==1.11
numpy==1.13.1
olefile==0.44
pbr==3.1.1
pexpect==4.0.1
Pillow==4.2.1
protobuf==3.3.0
psycopg2==2.7.3
ptyprocess==0.5
pycurl==7.43.0
PyMySQL==0.7.2
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==15.2.0
requests==2.11.1
scikit-image==0.13.0
scikit-learn==0.18.2
scipy==0.19.1
simplegeneric==0.8.1
six==1.10.0
stevedore==1.25.0
subprocess32==3.2.7
tensorflow==1.1.0
texttable==0.8.8
tornado==4.2.1
unity-lens-photos==1.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
websocket-client==0.44.0
Werkzeug==0.12.2
